+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 
 - Use a more mature (dog)statsd backend, with proper support for unix sockets and reservoir sampling.
   Due to the new backend, all sampling and tag filtering configuration options have been removed. ([#5675](https://github.com/getsentry/relay/pull/5675))
+- Remove unused `outcomes.emit_client_outcomes` configuration option. ([#5722](https://github.com/getsentry/relay/pull/5722))
 
 **Bug Fixes**:
 
+- Normalize custom metric units to `none` on trace metrics during processing. ([#5718](https://github.com/getsentry/relay/pull/5718))
 - Prevent minidump compression bomb. ([#5613](https://github.com/getsentry/relay/pull/5613))
 - Relay Temporarily rejected unknown headers on envelope items. ([#5709](https://github.com/getsentry/relay/pull/5709))
 
@@ -25,6 +27,7 @@
 
 - Strip performance metric specs from extraction while keeping extraction interfaces intact. ([#5674](https://github.com/getsentry/relay/pull/5674))
 - Allow deferred lengths to the `/upload` endpoint when the sender is trusted. ([#5658](https://github.com/getsentry/relay/pull/5658))
+- Stream non-prosperodmp attachments received at `/playstation` to objectstore. ([#5673](https://github.com/getsentry/relay/pull/5673))
 - Use new processor architecture to process client reports. ([#5686](https://github.com/getsentry/relay/pull/5686))
 - Use new processor architecture to process standalone attachments. ([#5703](https://github.com/getsentry/relay/pull/5703))
 - Prevent timeouts on the `/upload` endpoint. ([#5692](https://github.com/getsentry/relay/pull/5692))